+++ modules/ecs_cluster/cloudwatch_alarms.tf
@@ -1,0 +1,131 @@
+
+# ECS - CPU High
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "cluster_cpu_utilization_high" {
+  alarm_name          = "ecs-${var.name}-cluster-cpu-high"
+  comparison_operator = "GreaterThanThreshold"
+  evaluation_periods  = "2"
+  metric_name         = "CPUUtilization"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Average"
+  threshold           = var.cpu_threshold_high
+  alarm_description   = "ECS cluster CPU utilization is too high"
+  treat_missing_data  = "notBreaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}
+
+# ECS - Memory High
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "cluster_memory_utilization_high" {
+  alarm_name          = "ecs-${var.name}-cluster-memory-high"
+  comparison_operator = "GreaterThanThreshold"
+  evaluation_periods  = "2"
+  metric_name         = "MemoryUtilization"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Average"
+  threshold           = var.memory_threshold_high
+  alarm_description   = "ECS cluster memory utilization is too high"
+  treat_missing_data  = "notBreaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}
+
+# ECS - GPU High
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "gpu_utilization_high" {
+  alarm_name          = "ecs-${var.name}-cluster-gpu-high"
+  comparison_operator = "GreaterThanThreshold"
+  evaluation_periods  = "2"
+  metric_name         = "GPUUtilization"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Average"
+  threshold           = var.gpu_threshold_high
+  alarm_description   = "ECS cluster GPU utilization is too high"
+  treat_missing_data  = "notBreaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}
+
+# ECS - Tasks too low (cluster not used)
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "running_tasks_count_low" {
+  alarm_name          = "ecs-${var.name}-cluster-running-tasks-low"
+  comparison_operator = "LessThanThreshold"
+  evaluation_periods  = "2"
+  metric_name         = "RunningTasksCount"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Average"
+  threshold           = var.min_running_tasks_threshold
+  alarm_description   = "ECS cluster running tasks count is too low"
+  treat_missing_data  = "breaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}
+
+#####
+#####
+
+# ECS EC2 - Failed Tasks
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "task_count_failed" {
+  count               = var.type == "EC2" ? 1 : 0
+  alarm_name          = "ecs-${var.name}-cluster-failed-tasks"
+  comparison_operator = "GreaterThanThreshold"
+  evaluation_periods  = "1"
+  metric_name         = "FailedTaskCount"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Sum"
+  threshold           = var.failed_task_threshold
+  alarm_description   = "ECS cluster has failed tasks"
+  treat_missing_data  = "notBreaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}
+
+# ECS EC2 - Low Container Count
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_metric_alarm
+resource "aws_cloudwatch_metric_alarm" "container_instance_count_low" {
+  count               = var.type == "EC2" ? 1 : 0
+  alarm_name          = "ecs-${var.name}-cluster-container-instances-low"
+  comparison_operator = "LessThanThreshold"
+  evaluation_periods  = "2"
+  metric_name         = "RegisteredContainerInstancesCount"
+  namespace           = "AWS/ECS"
+  period              = "300"
+  statistic           = "Average"
+  threshold           = var.min_container_instances_threshold
+  alarm_description   = "ECS cluster container instance count is too low"
+  treat_missing_data  = "breaching"
+
+  dimensions = {
+    ClusterName = aws_ecs_cluster.main.name
+  }
+
+  tags = var.tags
+}