+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Edit: Refine VPC Flow Logs IAM Role/Policy
 - Edit: EC2 Monitoring enabled, EBS Optimization enabled, Dev Mode, remove AZ, Add Recovery Alarm
 - Edit: ASG EBS Optimized, Dynamic AMI Lookup, Encrypted Disk, Health Check, KMS Encryption, Dev Mode, Alarms
-- Edit: ECS Cluster to better utilize resources, Dev Mode, CloudWatch Log Group
+- Edit: ECS Cluster to better utilize resources, Dev Mode, CloudWatch Log Group, CloudWatch alarms
 - Edit: Load Balancer, more restrictive TLS/ciphers, cross-zone load balancing, CloudWatch alarms, encrypted WAF logs
 - Edit: WAF: rate limits, SQLi/Unix/Linux rule sets, Geo Blocking
 - Edit: SSM now uses its own KMS key, and has CloudTrail alarms